Convert Project to QT 5.6 + QTWebEngine. Affected module: Steam login. Workaround: Completed
Porting from QtWebkit to QTWebEngine was very annoying. QT 5.6 is
ridiculous, Worse cookie implementation, literally no HTTP-POST support.
The program now works but only after 8-9 hours of debugging.  Since
there is no POST support in the WebView (and we cannot just re-use a
NetworkAccessManager and shove it into the WebView like before) - the
Steam Login button takes you to the POE login area where you must
MANUALLY click the Steam Login button.  After realizing that the
POESESSID changes 3 times before "sync'ing" with the new STEAM
credentials, and developing a "gotcookie2 and 3" sentinel, the program
is working nearly as good as new, except for that extra click on "Login
with STEAM".